refactor: vip numbering
